refactor(loss_hub): move compute_ess_ratio_contribution to math_utils
ESS = (Σw)² / (n·Σw²) is a math primitive, sibling to compute_approx_kl /
compute_gspo_kl / compute_policy_loss. Moving it out of losses.py keeps
that file focused on the high-level per-loss-type dispatch and groups
the metric math next to the rest of the RL math toolkit.

Trim the corresponding losses.py imports: torch.distributed and
slice_loss_masks_for_local_cp are no longer needed there.
