+++ miles/backends/training_utils/loss_hub/losses.py
@@ -3,18 +3,17 @@ from collections.abc import Callable
 from typing import Protocol
 
 import torch
-import torch.distributed as dist
 
 from miles.backends.training_utils.cp_utils import (
     all_gather_with_cp,
     get_local_response_loss_masks,
     get_sum_of_sample_mean,
-    slice_loss_masks_for_local_cp,
 )
 from miles.backends.training_utils.loss_hub.corrections import vanilla_tis_function
 from miles.backends.training_utils.loss_hub.logit_processors import get_log_probs_and_entropy, get_values
 from miles.backends.training_utils.loss_hub.math_utils import (
     compute_approx_kl,
+    compute_ess_ratio_contribution,
     compute_gspo_kl,
     compute_opsm_mask,
     compute_policy_loss,
@@ -58,59 +57,6 @@ class LossFunction(Protocol):
                 on each implementation.
         """
         ...
-
-
-def compute_ess_ratio_contribution(
-    ppo_kl: torch.Tensor,
-    loss_masks: list[torch.Tensor],
-    total_lengths: list[int],
-    response_lengths: list[int],
-    qkv_format: str,
-    max_seq_lens: list[int] | None,
-    calculate_per_token_loss: bool,
-) -> torch.Tensor:
-    """Return an ESS contribution compatible with ``aggregate_train_losses``.
-
-    ESS needs full-sample sums before applying the nonlinear ratio.  Under CP we
-    reconstruct those sums first, then let only CP rank 0 emit the final
-    contribution so the generic CP metric aggregation does not double count it.
-    """
-    parallel_state = get_parallel_state()
-    cp_size = parallel_state.cp.size
-
-    local_masks = slice_loss_masks_for_local_cp(
-        loss_masks,
-        total_lengths,
-        response_lengths,
-        qkv_format,
-        max_seq_lens,
-    )
-    local_lengths = [mask.size(0) for mask in local_masks]
-    is_weights_per_sample = (-ppo_kl.detach().float()).exp().split(local_lengths, dim=0)
-
-    partial_sums = torch.zeros(len(loss_masks), 2, device=ppo_kl.device, dtype=torch.float32)
-    for i, (weights, mask) in enumerate(zip(is_weights_per_sample, local_masks, strict=False)):
-        if weights.numel() != mask.numel():
-            raise ValueError(f"ESS weight/mask length mismatch for sample {i}: {weights.numel()} vs {mask.numel()}")
-        masked_weights = weights * mask.to(device=weights.device, dtype=weights.dtype)
-        partial_sums[i, 0] = masked_weights.sum()
-        partial_sums[i, 1] = (masked_weights * masked_weights).sum()
-
-    if cp_size > 1:
-        dist.all_reduce(partial_sums, op=dist.ReduceOp.SUM, group=parallel_state.cp.group)
-
-    ess_ratio_sum = torch.zeros((), device=ppo_kl.device, dtype=torch.float32)
-    for i, loss_mask in enumerate(loss_masks):
-        num_valid_tokens = torch.clamp_min(loss_mask.to(device=ppo_kl.device, dtype=torch.float32).sum(), 1)
-        sum_w = partial_sums[i, 0]
-        sum_w2 = partial_sums[i, 1]
-        ess_ratio = (sum_w * sum_w) / (num_valid_tokens * torch.clamp_min(sum_w2, 1e-8))
-        ess_ratio_sum += ess_ratio * num_valid_tokens if calculate_per_token_loss else ess_ratio
-
-    if cp_size > 1 and parallel_state.cp.rank != 0:
-        ess_ratio_sum = ess_ratio_sum * 0
-
-    return ess_ratio_sum
 
 
 def policy_loss_function(
